requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
flask-swagger-ui==4.11.1
flasgger==0.9.7.1
PyMySQL==1.1.0
cryptography==41.0.7
bcrypt==4.1.2
python-dotenv==1.0.0
requests==2.31.0
redis==5.0.1
email-validator==2.1.0
marshmallow==3.20.1
werkzeug==3.0.1
sqlalchemy==2.0.25
numpy